Create utils_file
