+++ source/controller/controller.py
@@ -6,6 +6,7 @@ from source.controller.welcome_controller import WelcomeController
 from source.controller.wizard_controller import WizardController
 from source.controller.filter_controller import FilterController
 from source.controller.project_controller import ProjectController
+from source.domain.utils_file import import_gml_graph, create_gml_file
 from source.store.project_information_store import update_project_store
 from PyQt5.QtWidgets import *
 from source.store.project_information_store import project_information_store
@@ -13,7 +14,6 @@ import json
 from source.domain.exports import export_g6_to_png, export_g6_to_tikz, export_g6_to_pdf, export_g6_to_sheet
 from source.view.loading.loading_window import LoadingWindow
 from PyQt5 import QtCore
-from source.domain.utils import import_gml_graph, create_gml_file
 
 
 class Controller:
@@ -161,8 +161,10 @@ class Controller:
 
     def get_name_from_save_dialog(self, format_file):
         file_name = QFileDialog.getSaveFileName(parent=self.project_controller.project_window,
-                                                caption=self.project_controller.project_window.tr(f"Export graphs to {format_file} file"),
-                                                filter=self.project_controller.project_window.tr(f"Files (*.{format_file})"),
+                                                caption=self.project_controller.project_window.tr
+                                                (f"Export graphs to {format_file} file"),
+                                                filter=self.project_controller.project_window.tr
+                                                (f"Files (*.{format_file})"),
                                                 directory=f"{project_information_store.temp_project_name}.{format_file}"
                                                 )[0]
         if file_name:
@@ -186,7 +188,8 @@ class Controller:
     def export_to_png(self):
         self.project_controller.handle_tree_double_click()
         graph_to_export = self.get_graph_from_tree()
-        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window, caption="Select Directory"))
+        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window,
+                                                        caption="Select Directory"))
         if file_dir:
             self.show_loading_window(len(graph_to_export))
             for step, graph in enumerate(graph_to_export):
@@ -196,7 +199,8 @@ class Controller:
 
     def export_to_tikz(self):
         graph_to_export = self.get_graph_from_tree()
-        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window, caption="Select Directory"))
+        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window,
+                                                        caption="Select Directory"))
         if file_dir:
             self.show_loading_window(len(graph_to_export))
             for step, graph in enumerate(graph_to_export):
@@ -206,7 +210,8 @@ class Controller:
 
     def export_to_pdf(self):
         graph_to_export = self.get_graph_from_tree()
-        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window, caption="Select Directory"))
+        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window,
+                                                        caption="Select Directory"))
         if file_dir:
             self.show_loading_window(len(graph_to_export))
             for step, graph in enumerate(graph_to_export):
