+++ source/domain/utils.py
@@ -1,4 +1,3 @@
-import json
 import os
 import os.path
 import random
@@ -7,8 +6,6 @@ import networkx as nx
 import gzip
 
 from PyQt5.QtWidgets import QApplication
-
-from source.store.project_information_store import project_information_store
 
 
 def validate_path(path):
@@ -69,70 +66,6 @@ def set_view_size(self, p):
     self.height = int(rect.height() / p)
 
 
-def create_g6_file(file_path, g6):
-    try:
-        open(file_path, "x")
-    except FileExistsError:
-        pass
-    file = open(file_path, "w")
-    file.write(g6)
-    file.close()
-
-
-def change_g6_file(file_path, new_g6, current_index):
-    file = open(file_path, "r")
-    changed_data = file.readlines()
-
-    try:
-        changed_data[current_index] = new_g6 + "\n"
-    except IndexError:
-        changed_data.append(new_g6 + "\n")
-
-    with open(file_path, "w", encoding="utf-8") as file:
-        file.writelines(changed_data)
-
-    with open(file_path) as file:
-        graph = file.read().splitlines()
-
-    return graph
-
-
-def change_json_file(file_path, new_g6, current_index):
-    f = open(file_path)
-    data = json.load(f)
-    graph = list(data['filtered_graphs'])
-
-    try:
-        graph[current_index] = new_g6
-    except IndexError:
-        graph.append(new_g6)
-
-    graph = tuple(graph)
-    project_information_store.temp_filtered_graphs = graph
-    project_information_store.save_project()
-
-    f = open(file_path)
-    new_data = json.load(f)
-    new_json_graph = tuple(new_data['filtered_graphs'])
-
-    return new_json_graph
-
-
-def change_gml_file(file_path):
-    graph = project_information_store.current_graph
-    pos = project_information_store.current_graph_pos
-
-    for node, (x, y) in pos.items():
-        graph.nodes[node]['x'] = float(x)
-        graph.nodes[node]['y'] = float(y)
-
-    nx.write_gml(graph, file_path)
-
-
-def create_gml_file(graph, file_path):
-    nx.write_gml(graph, file_path)
-
-
 def fix_graph_nodes(graph):
     new_dict = {}
     new_edges = []
@@ -167,15 +100,3 @@ def set_new_vertex_positions(node_positions):
         new_y = random.uniform(0, 1)
 
     return new_x, new_y
-
-
-def import_gml_graph(file_path):
-    graph = nx.read_gml(file_path)
-
-    if len(nx.get_node_attributes(graph, 'x')) != 0:
-        for node in graph.nodes:
-            graph.nodes[node]['pos'] = (graph.nodes[node]['x'], graph.nodes[node]['y'])
-
-        project_information_store.current_graph_pos = nx.get_node_attributes(graph, 'pos')
-
-    return graph
